requirements.txt
python-dotenv>=1.0.0
google-cloud-vision>=3.0.0
requests>=2.28.0
pillow>=9.0.0
pillow-heif>=0.13.0
openai>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0

